third_party/WebKit/LayoutTests/fast/multicol/vertical-rl/offset-top-and-left-at-boundaries.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div style="position:relative; writing-mode:vertical-rl; columns:3; column-gap:10px; column-fill:auto; width:100px; height:320px;">
    <div id="first" style="width:100px;"></div>
    <div id="second" style="width:100px;"></div>
    <div id="third" style="width:100px;"></div>
</div>
<script>
    test(function() {
        assert_equals(document.getElementById('first').offsetLeft, 0);
        assert_equals(document.getElementById('first').offsetTop, 0);
    }, "offsetLeft and offsetTop in vertical-rl at column boundaries, first element");

    test(function() {
        assert_equals(document.getElementById('second').offsetLeft, 0);
        assert_equals(document.getElementById('second').offsetTop, 110);
    }, "offsetLeft and offsetTop in vertical-rl at column boundaries, second element");

    test(function() {
        assert_equals(document.getElementById('third').offsetLeft, 0);
        assert_equals(document.getElementById('third').offsetTop, 220);
    }, "offsetLeft and offsetTop in vertical-rl at column boundaries, third element");
</script>
